requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
drf-spectacular==0.22.1
environs==9.5.0
idna==3.3
importlib-resources==5.8.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.6.2
MarkupSafe==2.1.1
marshmallow==3.17.0
packaging==21.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.10
zipp==3.8.0
